Rewrite and reorganized the back-end Scrypts
Due to this project has been growing until becomes into a complex program it was necessary to reorganize the code separating the front-end and the back-end. This commit is the first step, rewrite and reorganize the old JGP1.py Scrypt into different Scrypts with cleaner functions and classes and better commentaries.

Apart of the reorganize, it was necessary to rethink the way that the classes and functions manage the data. For this reason the Variable object has been deleted and replaced with the use of two dictionaries, table and index.